+++ 06_dankovcev/03_alloc/my-allocator.h
@@ -2,9 +2,9 @@
 #include <stdio.h>
 #include <string.h>
 
-static const unsigned DEFAULT_MEM_REQUEST = 1 << 20;
+static const size_t DEFAULT_MEM_REQUEST = 1 << 20;
 
-void *my_malloc(unsigned space_amount);
+void *my_malloc(size_t space_amount);
 void my_free(void *ptr);
-void *my_calloc(unsigned space_amount);
-void *my_realloc(void *ptr, unsigned new_space_amount);
+void *my_calloc(size_t space_amount);
+void *my_realloc(void *ptr, size_t new_space_amount);
